Fixed naive backend, and added a few more examples                                                                                _ The stream interface is now composed of "int64 Stream.t" (instead of "int Stream.t" previously). The main node can now return several variables (see simple_double_out.prog for an example).
When the default value of fby is a constant, it can now be written either as a
tuple or as a integer (ie. "(0,1,0,1)" and "5" are identical in this context).

I also fixed the "convert" functions (that are generated to convert tuples to
other tuples with different sizes). They didn't work well, I only realised that
now.

And finally, I got rid of "Ocaml_runtime" module. (It was quite useless, only
exporting the identity for the boolean backed).