Adding generation of bitsliced code
- bitsliced (orthogonal) code can now be generated. The main differences with the
boolean code is the orthogonalisation of the data, the data types (int instead of
bool), and the fact that we need some kind of stack to comply with the stream
interface. (see for instance "src/tests/ocaml_run/simple_fby_ortho.ml")

- "src/tests/ocaml_run/test_fby.ml" has been updated with some new ideas about
fby. That's still a work in progress though.