Completed usuba implementation of DES
A complete implementation of DES (without mode, only on a 64bits input for now)
in Usuba is available in src/tests/des.prog.
The code naive generated by the compiler is in src/tests/ocaml_run/des_naive.prog
(the last 10 lines have been added to actually call the main on some data, but
every other line of the file have been generated).

I also added a "cache" to the "gen_conv" function (ocaml_gen_{naive,ortho}.ml),
thanks to which the same "convert" function can't be generated more than once
anymore.

Finally, I added the support for a "print" instruction. This isn't meant for
everyone to use, so it's not entirely convenient to use, but it is usefull for
debuging.