implemented fby on orthogonal backend
A new syntax has been introduce for fby : "x = v [f] fby d", where v is the
initial value, and f the function that will be used to expand v on 63 bits
(or rather on any number of bits, but for now we only generate orthogonal
code on 63 bits).
The syntax "x = v fby d" is still valid. In that case, the identity is used
to expand v on the others bits.
For instance, see src/tests/ortho_fby.prog. (and the generated OCaml code,
in "src/tests/ocaml_run/ortho_fby_ortho.ml").
This new syntax can be used even when generating naive code, but in that
case, "[f]" is ignored.
Note that there is still a little problem with this: the function generated
to expand "v" with "f" is declared too early in the generated code (in
particular, before "f"). This will be fixed soon.