Improved the architecture of the compiler
A lot of code was the same for the naive and orthogonal backend, but still was
duplicated. Furthermore, the whole compiler was in a single file, which was
starting to be a little bit too big.
Therefore, I changed the architecture, so there is now a single module "Rename"
that renames the identifier, a main module "Rewriter", that does most of the ast
transformation, and is parametrized by a module of type "SPECIFIC_REWRITER", that
will be specific to either the naive or orthogonal backend. And finally, a lot
of utility functions (join,indent,pow...) were also duplicated; they are now in
a separate module "Utils".