Added arrays and fill_i to usuba1
The main examples of what I've done in this commit are (src/tests/) :
 - array_node.prog for arrays of node
 - array_var.prog for arrays of variables
 - fill_i.prog for a basic example of fill
 - fill_advanced.prog for arrays of node, arrays of var and fill combined.
 - des_v1.prog to see how it can be used to simplify DES.

Arrays of nodes, and arrays of variables are mainly meant to simplify the
code, thanks to `fill_i` operator.
`a = fill_i<f;n>(init)` will fill the array `a` by repeatedly calling `f`,
with `init` as argument the first time, and `a[k-1]` at the `k`-th time.
Moreover, `fill_i` sends a variable to the node `f` (the "i" of fill_i),
which allows `f` to use arrays of node "dynamically" (actually, it's computed
at compile time), but it looks like `array[i]`.

All of those additions are part of Usuba1, which means that they are converted
to usuba0 code (by `src/usuba1_rewriter.ml`). Fundamentally, they are just
meant to make usuba code more pleasing to write.