Fix bug on arrays and rename usuba files
The extension for usuba files is now ".ua" (might change in the future
though).

The bug on array was that when an array was used without specifying
and index, it was considered as a boolean (?!) whereas one might
expect it to be equivalent to using every indexes at once.
For instance, let's say we had "arr:u2[2]". Then when we write for
instance, "arr = (1,1,2,1)", it is now compiled to something like
"(arr[0],arr[1]) = (1,1,2,1)" whereas it used to be left as is.