Add intrinsics in usuba, language verification and clean the tests
pand,por,pxor,pandn,etc. (see type "intr_op" in "usuba_AST") are now
valid Usuba. For now, the compiler converts "&" to "pand", "|" to por,
etc., but quite soon, the choice of which intrinsics to use will be
more subtle (and allowing the user to directly use intrinsics will
have a great importance on that choice).
Next steps are of course generation of C code, and then register
allocation, followed by the generation of a better C code.

The module "Assert_lang" has been added. Its role is to check wether
a program (Usuba_AST.prog) is valid in a given context. In particular,
whether it's valid usuba0, or it's normalized, or it only uses
intrinsics (and no "shift" or "arith" or "log" for instance). For now,
its purpose isn't crucial, but it will become more important later
(when we'll start thinking about the proofs).

Finally, I've cleaned and renamed the tests. Most of them weren't
working anyway, because they were using old syntaxes or old features.