Add benchmarks and fix tracking of variables
A lot of new benchmarks have been added. In particular, a comparison
between the different sizes of bitslicing (64, 128, 256) against
each other and against Kwan's implemented. See the script "compare.pl"
inside the directory "tests/perf_c".

Until now, the variables of every node weren't properly tracked. In
particular, every variable that is used should appear either in the
parameters or in the "vars" section. To compensate that, a function
"collect_decls" was used before generating the C code in order to
retrieve all variables used (because they need to be declared in C).
However, a crucial information was missing: their types. This was
becoming a problem because when choosing the intrinsics and the size
of the bitslicings (ie. choosing between MMX, SSE and AVX), we'd
like to know the variables types to generate appropriate
instructions.
Now that it's fixed, I'll soon implement different bitslicing. More
precisely, the C generation module will generate C code depending
on the variables types (whereas for now, it generates code for SSE
no matter what).