Update Z3 verification
The verification of the code with Z3 is now done in two passes.

The first one checks the equivalence between Usuba and Usuba0 no
inlined.
The second on checks the equivalence between Usuba0 no inlined and
Usuba0 inlined.

Try to check the equivalence between Usuba and Usuba0 at once was
too consuming and exponential (for 1-2-3-4 rounds of DES, it was
taking 1-3-50-70000 seconds...).

Now DES can be compiled and verified within roughly 30 seconds.