Update the compiler's structure / add arch modularity
The major reason for this modification is the addition of several
new backends (AltiVec and Neon are on the way), and the option
for the user to select which backend to use, with the flag:
 -arch <architecture>
where architecture can be any of {std,sse,avx,avx512,neon,altivec}

The compiler can now takes command line arguments (see the main
for a list of commands, or run "./usubac -h"). It allows the user
to specify various options about optimizations, backend and
configuration. This will make easier to generate and run tests
automatically.

The compiler now generates the runtime (whereas one used to need
to use the Perl script "gen_runtime.pl" to generate it).

Old unused files have been cleaned (everything related to the old
OCaml runtime, tests on old codes generated by previous versions
of the compiler, and unsued codes).