"flat values" semantics
The previous semantics was too complex and it actually diverged from
the implemented behavior.

The key idea is that "," must be seen as an associative operator on
tuples. For convenience, the singleton tuple "(x)" can be simply
written "x". As a result, everything is flat:

  ((x, y, z), (u, v), s, (d, e)) = (x, y, z, u, v, s, d, e)
  ((x, y, z), (u, v), s) << 1    = ((y, z), (u, v), s, 0)
                                 = (y, z, u, v, s, 0)

The semantics enforces this invariant by saying that a value is a list
of atoms, where an atom is a machine word of size 'n', suitable for
describing n-sliced algorithms.