More precise treatment of constants
Given an n : N, there are two ways to convert to a list of n_i's such
N.word_size n_i < atom_size.

We can either divide n by MAX_INT until the remainder is 0, at which
point we return the list (whose size therefore depends on the number
we started with). Or we set the size of the resulting list and divide
until a list of the desired size has been produced. If 'n' was too
small, this introduces padding by 0. If 'n' was too big, this drops
the most significant bits.

We chose the latter solution because
1. it is consistent with the current implementation
2. it is more predictable, the length of the tuple being forced by typing