Fix array expansion bug with main
The parameters of the main are now not expanded no matter what.
(we never want our main to have 200 parameters: arrays are always
better).
It used to be dealt with at the time of generating the C code, but
it was a bit hacky and make further optimizations (interleaving)
complicated.