Fix shift/rotate expansion
When writing (x << 5), x must be expanded regardless of
if we are keeping arrays or now: (x[0],..,x[10]) << 5 is
very different than (x << 5) (which actually means nothing).
To that end, the force parameter of expand_array.ml, used
to determine if arrays should be expanded or not as now 3
possible values instead of two: keep array (0), remove
arrays (1), split arrays (2).
The laster one (split arrays) means that we keep arrays,
but only as Index() and not as Var().