Change the type of variable declarations & add const keyword
This commit addresses the (long date) comment:
> (* XXX: define a record for [ident * typ * clock] *)
(from `src/usuba_ast.v`)

A type `var_d` of a single variable declaration has been added.
The type `p` is now a list of `var_d`.
And `var_d` is a reccord instead of a tuple of tuple.

At the same time, the keyword `const` was added, to specify that a
given stream is constant.
`var_d` thus have an attribute containing the attributes of the var.
Constant stream are needed when generating interleaved code: for
instance, the key is constant and thus shouldn't be duplicated.