Fix type of temporary variables
The motivation comes from `src/c_gen/nodes_to_c.ml`, line 103:
instead of an hardcoded "32", the size is now what it needs to be in
order for the shift/rotate to be correct (ie, it depends on the size
of the variable it's shifting).

To do so, the generation of temporary variables (in
`src/normalization/unfold_unnest.ml`) now generates variables with
correct types (instead of `Bool` and `Array(Bool,...)`).

A few implications:
 - Sboxes can't have general types `Bool[n] -> Bool[m]` because it
 messes up the types of the generated temporary variables. Therefore,
 Sboxes should be correctly typed. However, the Sbox database remains
 untyped (full boolean), and when loading one of its Sbox, the old
 type is kept (see `src/normalization/convert_tables.ml`).

 - We need to make sure that tuples of tuples are converted to tuples.
 This used to be done somewhere (don't remember where), but now it's
 more cleanly done in `src/normalization/norm_tuples.ml`.

Note: `src/optimization/share_var.ml` has been temporarily disabled
because it is bugged: in a loop, it can do something like replace
`x = a ^ b` with `a = a ^ b` even if `a` and `b` exist only inside
the loop while `x` exist outside (and should therefore not be
optimized out).