Don't overschedule for bitslicing
When doing bitslicing, the 'low_pressure_sched' scheduler will move
appart dependent variables, thus introducing spilling, which isn't
good. (and defeats the purpose of 'pre_schedule.ml')
(this is true mostly in the main functions; not in the Sboxes)

Therefore, 'low_pressure_sched' is used only if there are never more
than 60 variables live in the current function.
That 60 is a bit arbitrary, but bitsliced functions will always have
more (unless their blocksize is < 60, but I've never seen this), and
n-sliced/vectorized shouldn't have that much live variables.