Add lazy lifting option
This is useful for Rectangle: the key is a short, an it's broadcasted
when we need to use it.
It can be used on Serpent, but doesn't change the performances
(Kivilinna does this).

And will likely be useful for some other vectorized ciphers.

For now, it's not very powerful and needs to be activated with -lf
option.
A smarter version should check if we are actually doing vectorization,
and if so, choose to do it (or not).