Fix monomorphization of function parameters
Before, if `f: (in:b4[2]) ->...` was called as `f(x,...)` where `x:b8`,
then we'd end up with `f:b8` within monomorphize, which is obviously
wrong, since `in[0]` was a `b4` but is now a `b1` etc.

This is now fixed: only the direction 'D is updated by monomorphize.

Monomorphize should also update 'm when needed; I'll have to think
about that in more depth later, but I don't think this is an issue.