Add type-checker
Finally, the type-checker is back.

Type-checker is a bit missleading: "typer" might be better, since it
actually produces a new AST, enriched with some types: the types of
every expression and variables in Usuba can be deduced only using an
environement of variables (often called |env_var|) and an environment
of functions (often called |env_fun|). Everything but Const. The
type-checker thus performs type inference for Const, and reconstructs
and AST with typed Const.

The idea for now is twofold:

    - it types the Const, thus making Monomorphize and Unfold_unnest
      easier to implement (going to appear in future commits).

    - it should provide useful error messages when programing in
      Usuba.

There is one major limitation however: it doesn't deal with
polymorphism:

    - When an operator is applied to two operands, the typechecker for
      now only makes sure that both operand have the same
      type. Ideally, it should also make should that this operation
      can apply on operands of this type.
      I _think_ that this mostly apply to Shifts and Shuffles.
      (this is already done for Arith)

    - Furthermore, it should be the job of the type-checker to make
      sure that variables of different slicing direction don't
      interact (which it doesn't for now).

    - Finally, I would think that the type-checker is a good place to
      perform direction and word-size inference (which it doesn't do
      for now).