Fix expand_array bug with nested loops
The real bug was when two nested loops needs to be unrolled. For
instance,

    forall x in [1, 5] {
        forall y in [1, 5] {
            e(... x ... y ...)
        }
    }

if `forall x` needed unrolling, `forall y` would first be unrolled,
and during its expansion -which was bound to fail-, it would raise
Need_unroll, which would expand `forall x`. However, the iterator
environment was not cleaned, meaning that `forall x` would be unrolled
with an environment containing `y` as well (ie, it would replace `y`
with a value, which is obviously wrong.

This is now twice fixed:

  - Need_unroll precisely tells which loop needs unrolling

  - the iterator environement is now cleaned when Need_unrolled
    is raised.