Add Gift implementation
Note that the "standard" implementation (gift.ua) doesn't really work
in bitslicing. Or rather, it works only with -inline-all -unroll. This
is because temporary variables are not arrays in loops, which leads to
invalid code when later unrolling loops.

Thefore, there is a separate code, gift_bitslice.ua, which doesn't
suffer from this.