Remove clocks from Usuba
Clocks are never used in Usuba, and are not part of any
documentation. They were initially added because we thought it would
be useful to the language, but after 3 years, it turns out that we
never used clocks no needed them.

Keeping them around is doing more harm than good:

  - it makes some of the code needlessly heavier

  - anything that would require clocks to be working is probably wrong
    and will cause a lot of trouble to anyone trying to use them.

Therefore, it's better to remove them, and if at some point we think
that it would be a good addition, we will re-introduce them to Usuba.