Rework passes architecture
Running passes was a bit chaotic:

 - the normalizer was using a "pass_runner", but I was moderately
   convinced with how it behaved (registering all passes then running
   them).

  - the optimizer wasn't using the pass_runner...

  - there was verry little consistency between modules (debug messages
    weren't printed the same way for instance between the normalizer
    and the optimize).

  - debug messages were a bit hard to follow in nested passes (for
    instance, the Optimize was calling Simple_opts, which was calling
    Norm_tuples, which was calling Simplify_tuples, which is also
    called in plenty of other places, and it was very hard to find out
    in which call we actually were)

Now this should all be clearer. All passes have the same interface: a
variable |as_pass| which contains the name of the pass and the "run"
function (which has for signature pass_runner -> prog -> config -> prog).

I've renamed all entry points of all passes to "run" for more
concistency (and because it was needed for one of my failed
attempts...).