Fix bug in copy propagation
As things are currently, Tuples can survive very long in the pipeline
if they are in a Shift where the number to shift is a variable.
For instance,

    x = (a1, a2, a3) << (b * 2);

Can only be performed after inlining inlined the function if |b|
is a parameter, and after constant folding has replaced b*2 by
its actual value.