Rework inlining structure and pre-inlining condition
About the structure: when keeping all the inlining stuffs inside the
same file, I had an issue with a circular build (Shift_tuples requires
Inline, which requires Simple_opts, which eventually requires
Shift_tuples...). That's why Vital_inline is now in a separate file.

About the pre-inlining condition: pre-inlining is now a less
aggressive inlining, but with the same heuristic condition.