Functional hashtbl for the Ident module
Whenever a (ident, 'a) Hashtbl would be used, use a 'a Ident.Hashtbl
instead. This allows finer control over what hash and equality check
is used for identifiers

A new module, Prelude, has been added. This module monomorphize the
types of the Stdlib comparators to take integers and nothing else and
rejects the use of List.mem.